Implemented cookingpreset caching on online database. Get,Delete and Add are implemented. id in cookingpreset was modified from an int to a string to use the online database id generation. Methods that needed updating were changed accordingly.

+++ IoToaster App/IoToaster App/Services/CookingPresetService.cs
@@ -27,7 +27,7 @@ namespace IoToaster_App.Services
             await db.CreateTableAsync<CookingPreset>();
         }
 
-        public static async Task AddCookingPreset(string name, double toastDuration, int temperature)
+        public static async Task AddCookingPreset(string name, int toastDuration, int temperature)
         {
             await Init();
             var cookingPreset = new CookingPreset
